core/site.php: Improved handling of possible duplicate urls
Improved folderToUrl() and renamed it into makeUrl().
Now the function only transforms the actual slug of the file system path
and appends it to the parent URL.
If there are for example two parent folders "xxx.folder/" and "yyy.folder/",
they will be transformed into "folder/" and "folder-1/".
If the URL for "yyy.folder/child/" would be made from the whole file system path,
"folder/child/" would be returned by the function (wrong).
With appending only the transformed slug to the parent URL,
the returned URL will be "folder-1/child/" (correct).

To make this work, the parent URL is passed as an additional argument
for indexPagesRecursively(). The relative parent path got removed from the arguments.

The function filterSiteByParentRelPath() got removed.