Added core/selection.php
Introduced the Selection class to hold all methods for sorting and
filtering the collection of pages and return that modified
collection as "selection". That way the original $siteCollection from
the Site class stays untouched while filtering or sorting.

The filter methods no longer return a new filtered array, but
modify $this->selection directly. After all filter/sort is done, the
selection can be returned once by the getSelection() method.