core/page.php: First draft
A page object gets created in render.php and $_SERVER["PATH_INFO"]
gets passed as the parameter for the requested page.

The constructor of the page object then creates a new site object to
get the full site collection. All the current page's data gets stored
in $this->currentPage using the PATH_INFO as key for the
collection array.

Basically the Page class will handle the rendering of the
current page.