core/data.php: Added extractTags() method
Data::parseTxt() doesn't split the tags into an array anymore, but
keeps them as a normal $key/$value pair (string).

The new method Data::extractTags($data) looks for the tags-key and
returns an array of these tags.

Site::collectPages() uses this method now to make the tags a main
property of every page.