Added .htaccess
To make the RewriteBase "dynamic", the second RewriteCond is used
to get the correct path for the RewriteRule, while actually the
RewriteBase stays always in every setup just "/".

Very helpful information about solving portability this way can
be found in a tutorial by Michael Heim:
http://www.zeilenwechsel.de/it/articles/8/Using-mod_rewrite-in-.htaccess-files-without-knowing-the-RewriteBase.html

Basically it says:

We can't set the RewriteBase dynamically, so we set it once and for
all to the root URL of the server: RewriteBase /.
This provides consistency, but it also means that we have to establish
the url-path to the current directory ourselves and prefix it to the
rewritten URL.

So which directory are we in? Let's assume a REQUEST_URI
of /some/path/app-root/virtual/stuff. Our htaccess is in app-root.
If we grab the virtual part - virtual/stuff - and remove it from
the REQUEST_URI, we are left with the url-path to our app directory.

Capturing the virtual part is straightforward and can happen in the
rewrite rule itself. RewriteRule ^(.*)$ ... makes it available
in the $1 variable.

Now we do our little string operation and remove the virtual part
from the request URI. We don't have string commands for that,
but RewriteCond can match strings and capture substrings.
So we'll add a RewriteCond with the sole purpose of extracting
the url-path to the current directory. Other than that,
the "condition" should stay out of our way and always be true.

We can use the $1 variable from the RewriteRule in the RewriteCond
because mod_rewrite actually processes a ruleset backwards.
It starts with the pattern in the rule itself, and if it matches,
goes on to check the conditions. So the variable is available by then.

While the test string in the RewriteCond can use the variable,
the actual condition regex can't. Inside the condition, we can only
use internal back-references. So first, we assemble a test
string "[virtual part][some separator][request uri]".
The '#' char makes a good separator because it won't show up in the
URL. Next, we match it against a condition of

([^#]*) - anything up to the separator, captures the virtual part
#       - the separator
(.*?)   - anything in the request uri up to what we've captured in group one,
          grabs the current directory url-path
\1$     - group one again, ie the virtual part of the request uri

So here's the full condition:
RewriteCond $1#%{REQUEST_URI} ([^#]*)#(.*?)\1$.

The second captured group in the RewriteCond regex is our location.
We just need to prefix it to the rewritten URL with a %2 reference.
That leaves us with RewriteRule ^(.*)$ %2index.php [QSA,L].
Voil?.