core/template.php: Improved replacement of variables
Before $content was processed for each existing key in the data array.
Now $content gets processed only once and if there is a string found
between the variable delimiters, that string gets used as the key.

The replacement of the variables works now the same way like calling
the Tool functions.