core/template.php: Added modulateUrls()
Since all the generated links are now relative to Automad's root
directory, to make caching the templates easier and to improve the
overall portability, these links have to be adapted to Automad's
location.

To adapt these URLs (and also the relative URLs from the content
files) to the actual document root (in case Automad got installed in
a subfolder), modulateUrls() will parse the whole output and modulate
all URLs according to their type.