core/parse.php: Renamed sanitizeUrl() & updated core/site.php
Renamed sanitizeUrl() back to sanitize() to keep the naming consistent
when using the method for file names. It still does the same and has
nothing to do with the previously used sanitize() method
(used for stripping html tags).

The only place where the method is used at the moment is the Site
class, which got updated accordingly.