core/cache.php: The site's mTime gets now stored in a file
Before, it was actually possible that a page in the cache would never
get updated, since it was only one site scan every 30 seconds.
If a certain page would never hit that particular event and would be
always visited right after another page, it would be outdated forever.

Now the mTime of the site gets stored in a file. There is still a
delay due to performance issues, but that delays only the
re-scanning of all pages. So every 60 seconds the site gets
re-scanned. In between, the site's mTime gets loaded from a file.
So basically all pages should be up to date after these 60 seconds.

Additionally getLastestMTime() got renamed to getSiteMTime() and
$lastestMTime to $siteMTime due to the typos and for clarification.