core/extender.php: Simplified extension syntax: x{Name(Options)}
New file structure for extensions looks like this:
?extensions/name/name.php?

The main method that will be called from outside has now the same name
as the class itself.

An extension can now be called using x{extension(options)}. Since
class name and method name are equal, only the ?extension? name and
the options have to be specified.