Restructuring: core/gui.php: Handle session and includes within class
Since there are no single "GUI pages" anymore and the GUI gets initiated
in init.php, all session handling and including of the differnt GUI
modules (automad/gui/*) got moved into the class itself.

When a GUI object gets created, it gets first checked (in __construct())
whether an user is logged in. If an user is logged in, the "context"
(passed in the query string) defines, which module gets included.