core/modulate.php: Use page object as parameter instead of a page's ...
... path, to be able to also use its URL when creating relative URLs for
pages. Just using the request URI for a relative URL fails, when editing
a page in the GUI, since the requested URI is basically the GUI's URL
including a query string.

Note: It is still not recommended to use a relative redirect URL, since
the generated link to a page within any menu or navigation will always
be resolved relative to the currently viewed page and not to the actually
linked page.