Improved versioning scheme
Instead of generating major/minor release numbers automatically,
release.sh now just asks for a tag. Therefore it is possible to create
alpha and beta versions as well.
Since that also changes the format of the tags from X.Y to X.Y.Z,
to make sure the pre-release identifiers (alpha/beta) are always
appended to the patch version (Z), version.sh appends now a plus sign,
followed by the number of revisions ahead of the latest tag to
the latest tag to create a version number.